requirements.txt
altair==5.3.0
async-timeout==4.0.3
attrs==23.2.0
auctionsummary @ git+ssh://git@businessgit.ercot.com:7999/~dsherry/auction-summary.git@300331dc23cba859351e050976432010af03f61d
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
black==24.8.0
blinker==1.8.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cramjam==2.8.3
crontab==1.0.1
cryptography==42.0.5
cycler==0.12.1
entrypoints==0.4
ercotdb @ git+ssh://git@businessgit.ercot.com:7999/mav/ercotdb.git@d89b93c1f65eb8e90994d0ad36ff3f46551c056a
et-xmlfile==1.1.0
Faker==25.0.0
fastparquet==2024.5.0
favicon==0.7.0
fonttools==4.51.0
freezegun==1.5.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
htbuilder==0.6.2
idna==3.7
importlib_metadata==7.1.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==25.2.1
keyrings.alt==5.0.1
kiwisolver==1.4.5
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.8.2
mdurl==0.1.2
more-itertools==10.2.0
mypy-extensions==1.0.0
numpy==1.26.4
openpyxl==3.1.4
-e git+ssh://git@businessgit.ercot.com:7999/~dsherry/operatorinterface.git@49109cb72ded500988c56a4bcd8321e1186172e5#egg=operator
oracledb==2.3.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
plotly==5.18.0
polars==1.4.1
prometheus_client==0.20.0
protobuf==4.25.3
pyarrow==16.0.0
pycparser==2.22
pydeck==0.9.0
Pygments==2.17.2
pymdown-extensions==10.8.1
pyparsing==3.1.2
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
referencing==0.35.0
requests==2.31.0
requests-negotiate-sspi==0.5.2
rich==13.7.1
rpds-py==0.18.0
rq==1.16.2
schedule==1.2.2
selenium==4.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
st-annotated-text==4.0.1
streamlit==1.38.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.2
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-shadcn-ui==0.1.18
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tabulate==0.9.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
validators==0.28.1
watchdog==4.0.0
wsproto==1.2.0
wxPython==4.2.1
XlsxWriter==3.2.0
zipp==3.18.1
